moved 4b serialization tests to test_4bit
